Fixed misalignment Bug in EPG cache handling
This is more like papering over the real Problem. The problem
in eventData::get() is caused by allocating an unsigned int8 buffer
add an offset of 10 and access this as 32bit integer.

For the secondary fix the origin of the buffer is not that easy to discover.
This is hidden By something like the signal and slot concept in Qt. This was already
stated by "Dagobert" in the original comment.

This Patch is based on the work from TDT and sklnet.


Signed-off-by: Christian Ege <k4230r6@gmail.com>